feat: adding error and success tratment after submission
