+++ frontend/src/shared-constants/shared-constants.ts
@@ -1,4 +1,5 @@
 import MultiOptionsObj from '../types/MultiOptionsObject';
+import ResponseErrorType from '../types/ResponseErrorType';
 
 export const LOCATION_CODE_LIMIT = 2;
 export const SPAR_REDIRECT_PATH = 'SPAR-REDIRECT-PATH';
@@ -6,4 +7,10 @@ export const EmptyMultiOptObj: MultiOptionsObj = {
   code: '',
   label: '',
   description: ''
+};
+
+export const EmptyResponseError: ResponseErrorType = {
+  errOccured: false,
+  code: '',
+  message: ''
 };
