+++ frontend/src/types/ResponseErrorType.ts
@@ -1,0 +1,7 @@
+type ResponseErrorType = {
+  errOccured: boolean;
+  code: string;
+  message: string;
+};
+
+export default ResponseErrorType;
